Switch base image to node:alpine (#6)
Alpine is a lightweight Linux distribution designed for use in containers and aims to offer the required tools and packages at the minimum size.

The official Alpine-based Node.js image is 67MB, while the default Debian-based one is a whopping 672MB. Both images provide yarn by default, we just need to add git to be able to install `@hollowverse/common`.

This should result faster builds and deployments. I will submit PRs to switch the other repos to use Apline.