Set changeOrigin and toProxy options to true in proxy settings
I'm getting "Too Many Redirects" after I put Cloudfront in front of release manager on thehollowverse.com.

I'm not sure what's going on, but if I understand correctly, if `changeOrigin` is not set to `true`, the request will not have a `Host` header, and Cloudfront might be setting it back to the CF host.

I'll revert this if it causes any issues on hollowverse.com